Function libnotcurses_sys::ffi::ncvisual_media_defblitter

pub unsafe extern "C" fn ncvisual_media_defblitter(
    nc: *const notcurses, 
    scale: ncscale_e
) -> ncblitter_e

Get the default media (not plot) blitter for this environment when using the specified scaling method. Currently, this means:

if lacking UTF-8, NCBLIT_1x1

otherwise, if not NCSCALE_STRETCH, NCBLIT_2x1

otherwise, if sextants are not known to be good, NCBLIT_2x2

otherwise NCBLIT_3x2 NCBLIT_2x2 and NCBLIT_3x2 both distort the original aspect ratio, thus NCBLIT_2x1 is used outside of NCSCALE_STRETCH.