Struct libnotcurses_sys::ffi::ncplane

#[repr(C)]pub struct ncplane { /* fields omitted */ }

Implementations

impl ncplane

NcPlane Constructors

pub fn new<'a>(
nc: &mut Notcurses,
y: i32,
x: i32,
rows: u32,
cols: u32
) -> &'a mut NcPlane

New NcPlane.

The returned plane will be the top, bottom, and root of this new pile.

pub fn with_options<'a>(
nc: &mut Notcurses,
options: &NcPlaneOptions
) -> &'a mut NcPlane

New NcPlane, expects an NcPlaneOptions struct.

The returned plane will be the top, bottom, and root of this new pile.

pub fn new_bound<'a>(
bound_to: &mut NcPlane,
y: i32,
x: i32,
rows: u32,
cols: u32
) -> &'a mut NcPlane

New NcPlane, bound to another NcPlane.

pub fn with_options_bound<'a>(
nc: &mut Notcurses,
options: &NcPlaneOptions
) -> &'a mut NcPlane

New NcPlane, bound to another plane, expects an NcPlaneOptions struct.

The returned plane will be the top, bottom, and root of this new pile.

pub fn new_termsize<'a>(nc: &mut Notcurses) -> &'a mut NcPlane

New NcPlane, with the same dimensions of the terminal.

The returned plane will be the top, bottom, and root of this new pile.

impl ncplane

NcPlane Methods

pub fn cursor_yx(&self) -> (i32, i32)

Returns the current position of the cursor within the NcPlane.

Unlike [ncplane_cursor_yx] which uses i32, this uses u32.

pub fn cursor_y(&self) -> i32

Returns the current row of the cursor within the NcPlane.

pub fn cursor_x(&self) -> i32

Returns the current column of the cursor within the NcPlane.

pub fn dim_yx(&self) -> (u32, u32)

Return the dimensions of this NcPlane.

Unlike [ncplane_dim_yx] which uses i32, this uses u32.

pub fn dim_y(&self) -> u32

Return the rows of this NcPlane.

pub fn dim_x(&self) -> u32

Return the columns of this NcPlane.

pub fn set_scrolling(&mut self, scroll: bool) -> bool

Sets the scrolling behaviour of the plane, and returns true if scrolling was previously enabled, of false, if disabled.

All planes are created with scrolling disabled. Attempting to print past the end of a line will stop at the plane boundary, and indicate an error.

On a plane 10 columns wide and two rows high, printing "0123456789" at the origin should succeed, but printing "01234567890" will by default fail at the eleventh character. In either case, the cursor will be left at location 0x10; it must be moved before further printing can take place. I

pub fn erase(&mut self)

Erases every NcCell in the NcPlane, resetting all attributes to normal, all colors to the default color, and all cells to undrawn.

All cells associated with this NcPlane are invalidated, and must not be used after the call, excluding the base cell. The cursor is homed.

pub fn putc_yx(&mut self, y: i32, x: i32, cell: &NcCell) -> NcResult

Replaces the NcCell at the specified coordinates with the provided NcCell, advancing the cursor by its width (but not past the end of the plane).

The new NcCell must already be associated with the Plane. On success, returns the number of columns the cursor was advanced. On failure, -1 is returned.

pub fn putc(&mut self, cell: &NcCell) -> NcResult

Replaces the NcCell at the current coordinates with the provided NcCell, advancing the cursor by its width (but not past the end of the plane).

The new NcCell must already be associated with the Plane. On success, returns the number of columns the cursor was advanced. On failure, -1 is returned.

pub fn putstr(&mut self, string: &str) -> NcResult

Writes a series of [NcEgc]s to the current location, using the current style.

pub fn putstr_yx(&mut self, y: i32, x: i32, string: &str) -> NcResult

Write a string, which is a series of [NcEgc]s, to the current location, using the current style.

They will be interpreted as a series of columns (according to the definition of ncplane_putc()).

Advances the cursor by some positive number of columns (though not beyond the end of the plane); this number is returned on success.

On error, a non-positive number is returned, indicating the number of columns which were written before the error.

pub fn bottom<'a>(&mut self) -> &'a mut NcPlane

Returns the bottommost NcPlane of the pile that contains this [NnPlane].

pub fn top<'a>(&mut self) -> &'a mut NcPlane

Returns the topmost NcPlane of the pile that contains this [NnPlane].