Struct libnotcurses_sys::ffi::ncstats

#[repr(C)]
pub struct ncstats {

    pub renders: u64,
    pub writeouts: u64,
    pub failed_renders: u64,
    pub failed_writeouts: u64,
    pub render_bytes: u64,
    pub render_max_bytes: i64,
    pub render_min_bytes: i64,
    pub render_ns: u64,
    pub render_max_ns: i64,
    pub render_min_ns: i64,
    pub writeout_ns: u64,
    pub writeout_max_ns: i64,
    pub writeout_min_ns: i64,
    pub cellelisions: u64,
    pub cellemissions: u64,
    pub fgelisions: u64,
    pub fgemissions: u64,
    pub bgelisions: u64,
    pub bgemissions: u64,
    pub defaultelisions: u64,
    pub defaultemissions: u64,
    pub refreshes: u64,
    pub appsync_updates: u64,
    pub fbbytes: u64,
    pub planes: c_uint,
    pub raster_ns: u64,
    pub raster_max_ns: i64,
    pub raster_min_ns: i64,
    pub sprixelemissions: u64,
    pub sprixelelisions: u64,
    pub sprixelbytes: u64,
    pub input_errors: u64,
    pub input_events: u64,

}

whenever a new field is added here, ensure we add the proper rule to notcurses_stats_reset(), so that values are preserved in the stash stats.

Fields
renders: u64

successful ncpile_render() runs

writeouts: u64

successful ncpile_rasterize() runs

failed_renders: u64

aborted renders, should be 0

failed_writeouts: u64

aborted writes

render_bytes: u64

bytes emitted to ttyfp

render_max_bytes: i64

max bytes emitted for a frame

render_min_bytes: i64

min bytes emitted for a frame

render_ns: u64

nanoseconds spent rendering

render_max_ns: i64

max ns spent in render for a frame

render_min_ns: i64

min ns spent in render for a frame

writeout_ns: u64

nanoseconds spent writing frames to terminal

writeout_max_ns: i64

max ns spent writing out a frame

writeout_min_ns: i64

min ns spent writing out a frame

cellelisions: u64

cells we elided entirely thanks to damage maps

cellemissions: u64

total number of cells emitted to terminal

fgelisions: u64

RGB fg elision count

fgemissions: u64

RGB fg emissions

bgelisions: u64

RGB bg elision count

bgemissions: u64

RGB bg emissions

defaultelisions: u64

default color was emitted

defaultemissions: u64

default color was elided

refreshes: u64

refresh requests (non-optimized redraw)

appsync_updates: u64

how many application-synchronized updates?

fbbytes: u64

total bytes devoted to all active framebuffers

planes: c_uint

number of planes currently in existence

raster_ns: u64

nanoseconds spent rasterizing

raster_max_ns: i64

max ns spent in raster for a frame

raster_min_ns: i64

min ns spent in raster for a frame

sprixelemissions: u64

sprixel draw count

sprixelelisions: u64

sprixel elision count

sprixelbytes: u64

sprixel bytes emitted

input_errors: u64

errors processing control sequences/utf8

input_events: u64

characters returned to userspace

Implementations

impl ncstats

NcStats Methods.

pub fn new(nc: &mut Nc) -> &mut Self

Allocates an NcStats object.

pub fn stats(&mut self, nc: &mut Nc)

Acquires an atomic snapshot of the notcurses object’s stats.

pub fn reset(&mut self, nc: &mut Nc)

Resets all cumulative stats (immediate ones are not reset).