Struct libnotcurses_sys::ffi::nctree

#[repr(C)]
pub struct nctree { /* fields omitted */ }

Implementations

impl nctree

NcTree constructors & destructors

pub fn new<'a>(
plane: &mut NcPlane,
options: NcTreeOptions
) -> NcResult<&'a mut Self>

Creates an NcTree with the specified options.

C style function: nctree_create().

pub fn destroy(&mut self)

Destroys an NcTree created with new().

C style function: nctree_destroy().

impl nctree

NcTree methods

pub fn focused(&mut self) -> NcResult<&mut NcTreeItem>

Returns the focused item, if any items are present.

This is not a copy; be careful to use it only for the duration of a critical section.

C style function: nctree_focused().

pub fn next(&mut self) -> NcResult<&mut NcTreeItem>

Changes the focus to the next item, and returns it.

C style function: nctree_next().

pub fn prev(&mut self) -> NcResult<&mut NcTreeItem>

Changes the focus to the previous item, and returns it.

C style function: nctree_prev().

pub fn offer_input(&mut self, input: NcInput) -> bool

Offers the input to this NcTree.

If it’s relevant, this function returns true, and the input ought not be processed further. If it’s irrelevant to the tree, false is returned.

Relevant inputs include:

a mouse click on an item (focuses item)

a mouse scrollwheel event (srolls tree)

up, down, pgup, or pgdown (navigates among items)

C style function: nctree_offer_input().

pub fn plane(&mut self) -> NcResult<&NcPlane>

Returns the NcPlane backing this NcTree.

C style function: nctree_plane().

pub fn redraw(&mut self) -> NcResult<()>

Redraws the NcTree in its entirety.

The tree will be cleared, and items will be laid out, using the focused item as a fulcrum.

Item-drawing callbacks will be invoked for each visible item.

C style function: nctree_redraw().