Struct libnotcurses_sys::ffi::notcurses_options

#[repr(C)]
pub struct notcurses_options {
    pub termtype: *const c_char,
    pub renderfp: *mut FILE,
    pub loglevel: ncloglevel_e,
    pub margin_t: c_int,
    pub margin_r: c_int,
    pub margin_b: c_int,
    pub margin_l: c_int,
    pub flags: u64,
}

Configuration for notcurses_init().

Fields
termtype: *const c_char

The name of the terminfo database entry describing this terminal. If NULL, the environment variable TERM is used. Failure to open the terminal definition will result in failure to initialize notcurses.

renderfp: *mut FILE

If non-NULL, notcurses_render() will write each rendered frame to this FILE* in addition to outfp. This is used primarily for debugging.

loglevel: ncloglevel_e

Progressively higher log levels result in more logging to stderr. By default, nothing is printed to stderr once fullscreen service begins.

margin_t: c_int

Desirable margins. If all are 0 (default), we will render to the entirety of the screen. If the screen is too small, we do what we can–this is strictly best-effort. Absolute coordinates are relative to the rendering area ((0, 0) is always the origin of the rendering area).

margin_r: c_int

Desirable margins. If all are 0 (default), we will render to the entirety of the screen. If the screen is too small, we do what we can–this is strictly best-effort. Absolute coordinates are relative to the rendering area ((0, 0) is always the origin of the rendering area).

margin_b: c_int

Desirable margins. If all are 0 (default), we will render to the entirety of the screen. If the screen is too small, we do what we can–this is strictly best-effort. Absolute coordinates are relative to the rendering area ((0, 0) is always the origin of the rendering area).

margin_l: c_int

Desirable margins. If all are 0 (default), we will render to the entirety of the screen. If the screen is too small, we do what we can–this is strictly best-effort. Absolute coordinates are relative to the rendering area ((0, 0) is always the origin of the rendering area).

flags: u64

General flags; see NCOPTION_*. This is expressed as a bitfield so that future options can be added without reshaping the struct. Undefined bits must be set to 0.

Implementations

impl notcurses_options

NcOptions Constructors

pub const fn new() -> Self

New NcOptions.

pub const fn with_margins(
top: NcDim,
right: NcDim,
bottom: NcDim,
left: NcDim
) -> Self

New NcOptions, with margins.

pub const fn with_flags(flags: u64) -> Self

New NcOptions, with flags.

pub const fn with_all_options(
loglevel: NcLogLevel,
margin_t: NcDim,
margin_r: NcDim,
margin_b: NcDim,
margin_l: NcDim,
flags: u64
) -> Self

New NcOptions, with all the options.

Arguments

loglevel

Progressively higher log levels result in more logging to stderr. By default, nothing is printed to stderr once fullscreen service begins.

margin_t, margin_r, margin_b, margin_l

Desirable margins (top, right, bottom, left).

If all are 0 (default), we will render to the entirety of the screen. If the screen is too small, we do what we can. Absolute coordinates are relative to the rendering area ((0, 0) is always the origin of the rendering area).

flags

General flags; This is expressed as a bitfield so that future options can be added without reshaping the struct. Undefined bits must be set to 0.

NCOPTION_INHIBIT_SETLOCALE

NCOPTION_NO_ALTERNATE_SCREEN

NCOPTION_NO_FONT_CHANGES

NCOPTION_NO_QUIT_SIGHANDLERS

NCOPTION_NO_WINCH_SIGHANDLER

NCOPTION_SUPPRESS_BANNERS