Function libnotcurses_sys::nccell_prime

pub fn nccell_prime(
    plane: &mut NcPlane, 
    cell: &mut NcCell, 
    gcluster: &str, 
    style: NcStyle, 
    channels: NcChannels
) -> NcIntResult

Same as nccell_load, plus blasts the styling with style and channels.

Breaks the UTF-8 string in gcluster down, setting up the cell cell.

Returns the number of bytes copied out of gcluster, or -1 on failure.

The styling of the cell is left untouched, but any resources are released.

Blasts the styling with style and channels.

Method: NcCell.prime().