Function libnotcurses_sys::ncdirect_getc

pub unsafe extern "C" fn ncdirect_getc(
    n: *mut ncdirect, 
    ts: *const timespec, 
    sigmask: *mut sigset_t, 
    ni: *mut ncinput
) -> u32

See ppoll(2) for more detail. Provide a NULL ‘ts’ to block at length, a ‘ts’ of 0 for non-blocking operation, and otherwise a timespec to bound blocking. Signals in sigmask (less several we handle internally) will be atomically masked and unmasked per ppoll(2). ‘*sigmask’ should generally contain all signals. Returns a single Unicode code point, or (uint32_t)-1 on error. ‘sigmask’ may be NULL. Returns 0 on a timeout. If an event is processed, the return value is the ‘id’ field from that event. ‘ni’ may be NULL.