Function libnotcurses_sys::ncvisual_default_blitter

pub fn ncvisual_default_blitter(utf8: bool, scale: NcScale) -> NcBlitter

👎 Deprecated since 2.1.2:
please use notcurses_media_defblitter instead

Returns the best default blitter available

NCBLIT_3x2 is better image quality, especially for large images, but it's not the general default because it doesn't preserve aspect ratio. NCSCALE_STRETCH throws away aspect ratio, and can safely use NCBLIT_3x2.