Function libnotcurses_sys::ncvisual_stream

pub unsafe extern "C" fn ncvisual_stream(
    nc: *mut notcurses, 
    ncv: *mut ncvisual, 
    timescale: f32, 
    streamer: ncstreamcb, 
    vopts: *const ncvisual_options, 
    curry: *mut c_void
) -> c_int

Stream the entirety of the media, according to its own timing. Blocking, obviously. streamer may be NULL; it is otherwise called for each frame, and its return value handled as outlined for streamcb. If streamer() returns non-zero, the stream is aborted, and that value is returned. By convention, return a positive number to indicate intentional abort from within streamer(). 'timescale' allows the frame duration time to be scaled. For a visual naturally running at 30FPS, a 'timescale' of 0.1 will result in 300FPS, and a 'timescale' of 10 will result in 3FPS. It is an error to supply 'timescale' less than or equal to 0.