Macro libnotcurses_sys::prs

macro_rules! prs {
    ($p:expr, $( $sleep_args:expr),+ ) => { ... };
    ($nc:expr, $( $sleep_args:expr),+ ,) => { ... };
}

plane render sleep: NcPlane::render($p)? + NcPlane::rasterize($p)? + sleep![$sleep_args].

Renders and rasterizes the $p NcPlane pile and then, if there are no errors, calls the sleep macro with the rest of the arguments.

Returns NcResult.