Struct libnotcurses_sys::NcFile

pub struct NcFile { /* fields omitted */ }

A wrapper struct around libc::FILE

The notcurses FILE type FILE_NC a struct imported through bindgen, while the equivalent libc crate FILE (FILE_LIBC) is an opaque enum.

Several methods are provided to cast back and forth between both types, in order to allow both rust libc operations and notcurses file operations over the same underlying *FILE.

Implementations

impl NcFile

pub fn from_nc(file: *mut FILE_NC) -> Self

NcFile constructor from a file produced by notcurses

pub unsafe fn from_libc(file: *mut FILE_LIBC) -> Self

NcFile constructor from a file produced by the libc crate

pub fn as_libc_ptr(&self) -> *mut FILE_LIBC

Returns the file pointer in the format expected by the [libc] crate

pub fn as_nc_ptr(&self) -> *mut FILE_NC

Returns the file pointer in the format expected by notcurses

pub fn current_pos(&self) -> Result<u64, Error>

Returns the current position in the file.

On error Error::Errno(errno) is returned.

pub fn read_all(&mut self, buf: &mut Vec<u8>) -> Result<usize, Error>

Reads the file from start to end. Convenience method