Type Definition libnotcurses_sys::NcBlitter

type NcBlitter = ncblitter_e;

The blitter mode to use for rasterizing an NcVisual.

We never blit full blocks, but instead spaces (more efficient) with the background set to the desired foreground.

Modes

NCBLIT_1x1

NCBLIT_2x1

NCBLIT_2x2

NCBLIT_3x2

NCBLIT_4x1

NCBLIT_8x1

NCBLIT_BRAILLE

NCBLIT_DEFAULT

NCBLIT_PIXEL

There is a mechanism of graceful degradation, that works as follows:

without braille support, NCBLIT_BRAILLE decays to NCBLIT_3x2

without bitmap support, NCBLIT_PIXEL decays to NCBLIT_3x2

without sextant support, NCBLIT_3x2 decays to NCBLIT_2x2

without quadrant support, NCBLIT_2x2 decays to NCBLIT_2x1

the only viable blitters in ASCII are NCBLIT_1x1 and NCBLIT_PIXEL

If you don’t want this behaviour you have to use NCVISUAL_OPTION_NODEGRADE