Type Definition libnotcurses_sys::NcCell

type NcCell = cell;

A coordinate on an NcPlane storing 128 bits of data.

An NcCell corresponds to a single character cell on some NcPlane, which can be occupied by a single NcEgc grapheme cluster (some root spacing glyph, along with possible combining characters, which might span multiple columns).

At any NcCell, we can have a theoretically arbitrarily long UTF-8 string, a foreground color, a background color, and an NcStyleMask attribute set.

Valid grapheme cluster contents include:

A NUL terminator,

A single control character, followed by a NUL terminator,

At most one spacing character, followed by zero or more nonspacing characters, followed by a NUL terminator.

Diagram

NcCell: 128 bits structure comprised of the following 5 elements:

GCLUSTER GCLUSTER GCLUSTER GCLUSTER  1. NcEgc
00000000 ~~~~~~~~ 11111111 11111111  2. NcEgcBackstop + 3. width + 4. NcStyleMask
~~AA~~~~ RRRRRRRR GGGGGGGG BBBBBBBB  5. NcChannelPair
~~AA~~~~ RRRRRRRR GGGGGGGG BBBBBBBB  |

1. (32b) Extended Grapheme Cluster, presented either as:

    1.1. An NcEgc of up to 4 bytes:
    UUUUUUUU UUUUUUUU UUUUUUUU UUUUUUUU

    1.2. A `0x01` in the first byte, plus 3 bytes with a 24b address to an egcpool:
    00000001 IIIIIIII IIIIIIII IIIIIIII

2. (8b) Backstop (zero)
00000000

3. (8b) width (biased 1)
~~~~~~~~

4. (16b) NcStyleMask
11111111 11111111

5. (64b) NcChannelPair
~~AA~~~~ RRRRRRRR GGGGGGGG BBBBBBBB|~~AA~~~~ RRRRRRRR GGGGGGGG BBBBBBBB

type in C: cell (struct)

Size

Multi-column characters can only have a single style/color throughout. wcwidth() is not reliable. It's just quoting whether or not the NcEgc contains a "Wide Asian" double-width character. This is set for some things, like most emoji, and not set for other things, like cuneiform.

Each cell occupies 16 static bytes (128 bits). The surface is thus ~1.6MB for a (pretty large) 500x200 terminal. At 80x43, it's less than 64KB. Dynamic requirements (the egcpool) can add up to 16MB to an ncplane, but such large pools are unlikely in common use.

Alpha Compositing

We implement some small alpha compositing. Foreground and background both have two bits of inverted alpha. The actual grapheme written to a cell is the topmost non-zero grapheme.

If its alpha is 00 (NCCELL_ALPHA_OPAQUE) its foreground color is used unchanged.

If its alpha is 10 (NCCELL_ALPHA_TRANSPARENT) its foreground color is derived entirely from cells underneath it.

If its alpha is 01 (NCCELL_ALPHA_BLEND) the result will be a composite.

Likewise for the background. If the bottom of a coordinate's zbuffer is reached with a cumulative alpha of zero, the default is used. In this way, a terminal configured with transparent background can be supported through multiple occluding ncplanes.

A foreground alpha of 11 (NCCELL_ALPHA_HIGHCONTRAST) requests high-contrast text (relative to the computed background). A background alpha of 11 is currently forbidden.

Precedence

Default color takes precedence over palette or RGB, and cannot be used with transparency.

Indexed palette takes precedence over RGB. It cannot meaningfully set transparency, but it can be mixed into a cascading color.

RGB is used if neither default terminal colors nor palette indexing are in play, and fully supports all transparency options.

Implementations

impl NcCell

NcCell Constructors

pub const fn with_char7b(ch: char) -> Self

New NcCell, expects a 7-bit char.

pub fn with_char(ch: char, plane: &mut NcPlane) -> Self

New NcCell, expects an NcPlane and a char.

pub fn with_str(plane: &mut NcPlane, string: &str) -> Self

New NcCell, expects an NcPlane and a &str.

pub const fn new() -> Self

New empty NcCell.

impl NcCell

NcPlane methods: bg|fg NcChannels manipulation.

pub fn bchannel(&self) -> NcChannel

Gets the background NcChannel.

C style function: cell_bchannel().

pub fn bg_alpha(&self) -> NcAlphaBits

Extracts the background NcAlphaBits (shifted to LSBs).

C style function: cell_bg_alpha().

pub fn bg_default_p(&self) -> bool

Is the background NcChannel using the "default background color"?

C style function: cell_bg_default_p().

pub fn bg_palindex(&self) -> NcPaletteIndex

Gets the NcPaletteIndex of the background NcChannel.

C style function: cell_bg_palindex().

pub fn bg_palindex_p(&self) -> bool

Is the background NcChannel using an NcPaletteIndex indexed NcPalette color?

C style function: cell_bg_palindex_p().

pub fn bg_rgb(&self) -> NcRgb

Gets the background NcRgb (shifted to LSBs).

C style function: cell_bg_rgb().

pub fn bg_rgb8(&self) -> (NcColor, NcColor, NcColor)

Gets the background NcColor RGB components.

C style function: cell_bg_rgb8().

pub fn fchannel(&self) -> NcChannel

Gets the foreground NcChannel.

C style function: cell_fchannel().

pub fn fg_alpha(&self) -> NcAlphaBits

Extracts the foreground NcAlphaBits (shifted to LSBs).

C style function: cell_fg_alpha().

pub fn fg_default_p(&self) -> bool

Is the foreground NcChannel using the "default foreground color"?

C style function: cell_fg_default_p().

pub fn fg_palindex(&self) -> NcPaletteIndex

Gets the NcPaletteIndex of the foreground NcChannel.

C style function: cell_fg_palindex().

pub fn fg_palindex_p(&self) -> bool

Is the foreground NcChannel using an NcPaletteIndex indexed NcPalette color?

C style function: cell_fg_palindex_p().

pub fn fg_rgb(&self) -> NcRgb

Gets the foreground NcRgb (shifted to LSBs).

C style function: cell_fg_rgb().

pub fn fg_rgb8(&self) -> (NcColor, NcColor, NcColor)

Gets the foreground NcColor RGB components.

C style function: cell_fg_rgb8().

pub fn set_bchannel(&mut self, channel: NcChannel) -> NcChannelPair

Sets the background NcChannel and returns the new NcChannelPair.

C style function: cell_set_bchannel().

pub fn set_bg_alpha(&mut self, alpha: NcAlphaBits)

Sets the background NcAlphaBits.

C style function: cell_set_bg_alpha().

pub fn set_bg_default(&mut self)

Indicates to use the "default color" for the background NcChannel.

C style function: cell_set_bg_default().

pub fn set_bg_palindex(&mut self, index: NcPaletteIndex)

Sets the background NcPaletteIndex.

Also sets NCCELL_BG_PALETTE and NCCELL_ALPHA_OPAQUE, and clears out NCCELL_BGDEFAULT_MASK.

C style function: cell_set_bg_palindex().

pub fn set_bg_rgb(&mut self, rgb: NcRgb)

Sets the background NcRgb and marks it as not using the default color.

C style function: cell_set_bg_rgb().

pub fn set_bg_rgb8(&mut self, red: NcColor, green: NcColor, blue: NcColor)

Sets the background NcColor RGB components, and marks it as not using the "default color".

C style function: cell_set_bg_rgb8().

pub fn set_fchannel(&mut self, channel: NcChannel) -> NcChannelPair

Sets the foreground NcChannel and returns the new NcChannelPair.

C style function: cell_set_fchannel().

pub fn set_fg_alpha(&mut self, alpha: NcAlphaBits)

Sets the foreground NcAlphaBits.

C style function: cell_set_fg_alpha().

pub fn set_fg_default(&mut self)

Indicates to use the "default color" for the foreground NcChannel.

C style function: cell_set_fg_default().

pub fn set_fg_palindex(&mut self, index: NcPaletteIndex)

Sets the foreground NcPaletteIndex.

Also sets NCCELL_FG_PALETTE and NCCELL_ALPHA_OPAQUE, and clears out NCCELL_BGDEFAULT_MASK.

C style function: cell_set_fg_palindex().

pub fn set_fg_rgb(&mut self, rgb: NcRgb)

Sets the foreground NcRgb and marks it as not using the default color.

C style function: cell_set_fg_rgb().

pub fn set_fg_rgb8(&mut self, red: NcColor, green: NcColor, blue: NcColor)

Sets the foreground NcColor RGB components, and marks it as not using the "default color".

C style function: cell_set_fg_rgb8().

impl NcCell

NcCell Methods

pub fn extract(
&mut self,
plane: &mut NcPlane,
styles: &mut NcStyleMask,
channels: &mut NcChannelPair
) -> NcEgc

Saves the NcStyleMask and the NcChannelPair, and returns the NcEgc (the three elements of an NcCell).

C style function: cell_fg_alpha().

pub fn channels(&mut self, plane: &mut NcPlane) -> NcChannelPair

Returns the NcChannelPair of the NcCell.

(No equivalent C style function)

pub fn styles(&mut self, plane: &mut NcPlane) -> NcStyleMask

Returns the NcStyleMask of the NcCell.

(No equivalent C style function)

pub fn egc(&mut self, plane: &mut NcPlane) -> NcEgc

Returns the NcEgc of the NcCell.

(No equivalent C style function)