Type Definition libnotcurses_sys::NcChannelPair

type NcChannelPair = u64;

64 bits containing a foreground and background NcChannel

At render time, both 24-bit NcRgb values are quantized down to terminal capabilities, if necessary. There's a clear path to 10-bit support should we one day need it.

Default Color

The "default color" is best explained by color(3NCURSES) and default_colors(3NCURSES). Ours is the same concept.

Until the "not default color" bit is set, any color you load will be ignored.

Diagram

~~AA~~~~|RRRRRRRR|GGGGGGGG|BBBBBBBB|~~AA~~~~|RRRRRRRR|GGGGGGGG|BBBBBBBB
↑↑↑↑↑↑↑↑↑↑↑↑ foreground ↑↑↑↑↑↑↑↑↑↑↑|↑↑↑↑↑↑↑↑↑↑↑↑ background ↑↑↑↑↑↑↑↑↑↑↑

Detailed info (specially on the context-dependent bits on each NcChannel's 4th byte):

                            ~foreground channel~
NCCELL_WIDEASIAN_MASK: part of a wide glyph          ↓bits view↓               ↓hex mask↓
1······· ········ ········ ········ ········ ········ ········ ········  =  8······· ········

NCCELL_FGDEFAULT_MASK: foreground is NOT "default color"
·1······ ········ ········ ········ ········ ········ ········ ········  =  4······· ········

NCCELL_FG_ALPHA_MASK: foreground alpha (2bits)
··11···· ········ ········ ········ ········ ········ ········ ········  =  3······· ········

NCCELL_FG_PALETTE: foreground uses palette index
····1··· ········ ········ ········ ········ ········ ········ ········  =  ·8······ ········

NCCELL_NOBACKGROUND_MASK: glyph is entirely foreground
·····1·· ········ ········ ········ ········ ········ ········ ········  =  ·4······ ········

reserved, must be 0
······00 ········ ········ ········ ········ ········ ········ ········  =  ·3······ ········

NCCELL_FG_RGB_MASK: foreground in 3x8 RGB (rrggbb)
········ 11111111 11111111 11111111 ········ ········ ········ ········  =  ··FFFFFF ········

                            ~background channel~
reserved, must be 0                                  ↓bits view↓               ↓hex mask↓
········ ········ ········ ········ 0······· ········ ········ ········  =  ········ 8·······

NCCELL_BGDEFAULT_MASK: background is NOT "default color"
········ ········ ········ ········ ·1······ ········ ········ ········  =  ········ 4·······

NCCELL_BG_ALPHA_MASK: background alpha (2 bits)
········ ········ ········ ········ ··11···· ········ ········ ········  =  ········ 3·······

NCCELL_BG_PALETTE: background uses palette index
········ ········ ········ ········ ····1··· ········ ········ ········  =  ········ ·8······

reserved, must be 0
········ ········ ········ ········ ·····000 ········ ········ ········  =  ········ ·7······

NCCELL_BG_RGB_MASK: background in 3x8 RGB (rrggbb)
0········ ········ ········ ········ ········11111111 11111111 11111111  =  ········ ··FFFFFF

type in C: channels (uint64_t)

NcCell Mask Flags

NCCELL_BGDEFAULT_MASK

NCCELL_BG_ALPHA_MASK

NCCELL_BG_PALETTE

NCCELL_BG_RGB_MASK

NCCELL_FGDEFAULT_MASK

NCCELL_FG_ALPHA_MASK

NCCELL_FG_PALETTE

NCCELL_FG_RGB_MASK