Type Definition libnotcurses_sys::NcEgc

type NcEgc = char;

Extended Grapheme Cluster. A 32-bit char-like type

This 32 bit char, together with the associated plane's associated egcpool, completely define this cell's NcEgc. Unless the NcEgc requires more than four bytes to encode as UTF-8, it will be inlined here:

Diagram 1

UUUUUUUU UUUUUUUU UUUUUUUU UUUUUUUU
extended grapheme cluster <= 4bytes

type in C: uint32_t

If more than four bytes are required, it will be spilled into the egcpool. In either case, there's a NUL-terminated string available without copying, because (1) the egcpool is all NUL-terminated sequences and (2) the fifth byte of this struct (the GClusterBackStop field, see below) is guaranteed to be zero, as are any unused bytes in gcluster.

A spilled NcEgc is indicated by the value 0x01iiiiii. This cannot alias a true supra-ASCII NcEgc, because UTF-8 only encodes bytes <= 0x80 when they are single-byte ASCII-derived values. The iiiiii is interpreted as a 24-bit index into the egcpool (which may thus be up to 16MB):

Diagram 2

00000001 iiiiiiii iiiiiiii iiiiiiii
  sign     24bit index to egcpool

type in C: uint32_t

The cost of this scheme is that the character 0x01 (SOH) cannot be encoded in a cell, and therefore it must not be allowed through the API.

NOTE that even if the NcEgc is <= 4 bytes and inlined, is still interpreted as a NUL-terminated char * (technically, &cell->gcluster is treated as a char*). If it is more than 4 bytes, cell->gcluster has a first byte of 0x01, and the remaining 24 bits are an index into the plane's egcpool, which is carved into NUL-terminated chunks of arbitrary length.

Links

Grapheme Cluster Boundaries