Type Definition libnotcurses_sys::NcInput

type NcInput = ncinput;

Reads and decodes input events

Reads from stdin and decodes the input to stdout, including synthesized events and mouse events.

Notcurses provides input from keyboards and mice. Single Unicode codepoints are received from the keyboard, directly encoded as u32.

The input system must deal with numerous keyboard signals which do not map to Unicode code points. This includes the keypad arrows and function keys. These “synthesized” codepoints are enumerated in , and mapped into the Supplementary Private Use Area-B (U+100000..U+10FFFD). Mouse button events are similarly mapped into the SPUA-B.

All events carry a ncinput structure with them. For mouse events, the x and y coordinates are reported within this struct. For all events, modifiers (e.g. “Alt”) are carried as bools in this struct.

Implementations

impl NcInput

New NcInput.

pub const fn new_empty() -> NcInput

New empty NcInput.

pub const fn new(id: char) -> NcInput

New NcInput.

pub const fn with_alt(id: char) -> NcInput

New NcInput with alt key.

pub const fn with_shift(id: char) -> NcInput

New NcInput with shift key.

pub const fn with_ctrl(id: char) -> NcInput

New NcInput with ctrl key.

pub const fn with_all_args(
id: char,
x: Option<NcDim>,
y: Option<NcDim>,
alt: bool,
shift: bool,
ctrl: bool,
seqnum: u64
) -> NcInput

New NcInput, expecting all the arguments.