Type Definition libnotcurses_sys::NcMenu

type NcMenu = ncmenu;

menus on the top or bottom rows

A notcurses instance supports menu bars on the top or bottom row of the true screen.

A menu is composed of sections, which are in turn composed of items. Either no sections are visible, and the menu is rolled up, or exactly one section is unrolled.

ncmenu_rollup places an NcMenu in the rolled up state. ncmenu_unroll rolls up any unrolled section and unrolls the specified one. ncmenu_destroy removes a menu bar, and frees all associated resources.

type in C: ncmenu (struct)

Implementations

impl NcMenu

NcMenu Constructors

pub unsafe fn new<'a>(plane: &mut NcPlane) -> &'a mut Self

NcMenu simple constructor

pub unsafe fn with_options<'a>(
plane: &mut NcPlane,
options: &NcMenuOptions
) -> &'a mut Self

NcMenu constructor with options