Type Definition libnotcurses_sys::NcPlane

type NcPlane = ncplane;

Fundamental drawing surface.

Unites a:

CellMatrix

EgcPool

type in C: ncplane (struct)

Piles

A single notcurses context is made up of one or more piles.

A pile is a set of one or more ncplanes, including the partial orderings made up of their binding and z-axis pointers.

A pile has a top and bottom ncplane (this might be a single plane), and one or more root planes (planes which are bound to themselves).

Multiple threads can concurrently operate on distinct piles, even changing one while rendering another.

Each plane is part of one and only one pile. By default, a plane is part of the same pile containing that plane to which it is bound.

If ncpile_create is used in the place of ncplane_create, the returned plane becomes the root plane, top, and bottom of a new pile. As a root plane, it is bound to itself.

A new pile can also be created by reparenting a plane to itself, though if the plane is already a root plane, this is a no-op.

When a plane is moved to a different pile (whether new or preexisting), any planes which were bound to it are rebound to its previous parent. If the plane was a root plane of some pile, any bound planes become root planes. The new plane is placed immediately atop its new parent on its new pile's z-axis. When ncplane_reparent_family() is used, all planes bound to the reparented plane are moved along with it. Their relative z-order is maintained.

Implementations

impl NcPlane

NcPlane constructors and destructors

pub fn new<'a>(
nc: &mut Notcurses,
y: i32,
x: i32,
rows: u32,
cols: u32
) -> &'a mut NcPlane

New NcPlane.

The returned plane will be the top, bottom, and root of this new pile.

pub fn with_options<'a>(
nc: &mut Notcurses,
options: &NcPlaneOptions
) -> &'a mut NcPlane

New NcPlane, expects an NcPlaneOptions struct.

The returned plane will be the top, bottom, and root of this new pile.

pub fn new_bound<'a>(
bound_to: &mut NcPlane,
y: i32,
x: i32,
rows: u32,
cols: u32
) -> &'a mut NcPlane

New NcPlane, bound to another NcPlane.

pub fn with_options_bound<'a>(
nc: &mut Notcurses,
options: &NcPlaneOptions
) -> &'a mut NcPlane

New NcPlane, bound to another plane, expects an NcPlaneOptions struct.

The returned plane will be the top, bottom, and root of this new pile.

pub fn new_termsize<'a>(nc: &mut Notcurses) -> &'a mut NcPlane

New NcPlane, with the same dimensions of the terminal.

The returned plane will be the top, bottom, and root of this new pile.

pub fn destroy(&mut self) -> NcResult

Destroys the NcPlane. None of its contents will be visible after the next render call. It is an error to attempt to destroy the standard plane.

impl NcPlane

NcPlane Methods

pub fn cursor_yx(&self) -> (i32, i32)

Returns the current position of the cursor within the NcPlane.

Unlike ncplane_cursor_yx which uses i32, this uses u32.

pub fn cursor_y(&self) -> i32

Returns the current row of the cursor within the NcPlane.

pub fn cursor_x(&self) -> i32

Returns the current column of the cursor within the NcPlane.

pub fn cursor_move_yx(&mut self, y: i32, x: i32) -> NcResult

Moves the cursor to the specified position within the NcPlane.

The cursor doesn't need to be visible.

Negative parameters, or ones exceeding the plane's dimensions will result in an error, and the cursor position will remain unchanged.

pub fn cursor_move_rows(&mut self, rows: i32) -> NcResult

Moves the cursor the number of rows specified (forward or backwards).

It will error if the target row exceeds the plane dimensions.

pub fn cursor_move_cols(&mut self, cols: i32) -> NcResult

Moves the cursor the number of columns specified (forward or backwards).

It will error if the target column exceeds the plane dimensions.

pub fn dim_yx(&self) -> (u32, u32)

Return the dimensions of this NcPlane.

Unlike ncplane_dim_yx which uses i32, this uses u32.

pub fn dim_y(&self) -> u32

Return the rows of this NcPlane.

pub fn dim_x(&self) -> u32

Return the columns of this NcPlane.

pub fn rows(&self) -> u32

Return the rows of this NcPlane.

pub fn cols(&self) -> u32

Return the cols of this NcPlane.

pub fn set_scrolling(&mut self, scroll: bool) -> bool

Sets the scrolling behaviour of the plane, and returns true if scrolling was previously enabled, of false, if disabled.

All planes are created with scrolling disabled. Attempting to print past the end of a line will stop at the plane boundary, and indicate an error.

On a plane 10 columns wide and two rows high, printing "0123456789" at the origin should succeed, but printing "01234567890" will by default fail at the eleventh character. In either case, the cursor will be left at location 0x10; it must be moved before further printing can take place. I

pub fn erase(&mut self)

Erases every NcCell in the NcPlane, resetting all attributes to normal, all colors to the default color, and all cells to undrawn.

All cells associated with this NcPlane are invalidated, and must not be used after the call, excluding the base cell. The cursor is homed.

pub fn putc_yx(&mut self, y: i32, x: i32, cell: &NcCell) -> NcResult

Replaces the NcCell at the specified coordinates with the provided NcCell, advancing the cursor by its width (but not past the end of the plane).

The new NcCell must already be associated with the Plane. On success, returns the number of columns the cursor was advanced. On failure, -1 is returned.

pub fn putc(&mut self, cell: &NcCell) -> NcResult

Replaces the NcCell at the current coordinates with the provided NcCell, advancing the cursor by its width (but not past the end of the plane).

The new NcCell must already be associated with the Plane. On success, returns the number of columns the cursor was advanced.

pub fn putchar(&mut self, ch: char) -> NcResult

Calls ncplane_putchar_yx() at the current cursor location.

pub fn putchar_yx(&mut self, y: i32, x: i32, ch: char) -> NcResult

Replaces the NcEgc, but retain the styling. The current styling of the plane will not be changed.

pub fn putstr(&mut self, string: &str) -> NcResult

Writes a series of NcEgcs to the current location, using the current style.

Advances the cursor by some positive number of columns (though not beyond the end of the plane); this number is returned on success.

On error, a non-positive number is returned, indicating the number of columns which were written before the error.

pub fn putstr_stained(&mut self, string: &str) -> NcResult

Writes a series of NcEgcs to the current location, but retain the styling. The current styling of the plane will not be changed.

pub fn putstr_yx(&mut self, y: i32, x: i32, string: &str) -> NcResult

Write a string, which is a series of NcEgcs, to the current location, using the current style.

They will be interpreted as a series of columns (according to the definition of ncplane_putc()).

Advances the cursor by some positive number of columns (though not beyond the end of the plane); this number is returned on success.

On error, a non-positive number is returned, indicating the number of columns which were written before the error.

pub fn bottom<'a>(&mut self) -> &'a mut NcPlane

Returns the bottommost NcPlane of the pile that contains this NcPlane.

pub fn top<'a>(&mut self) -> &'a mut NcPlane

Returns the topmost NcPlane of the pile that contains this NcPlane.

pub fn move_yx(&mut self, y: i32, x: i32) -> NcResult

Move this NcPlane relative to the standard plane, or the plane to which it is bound. It is an error to attempt to move the standard plane.

pub fn notcurses<'a>(&mut self) -> &'a mut Notcurses

Gets a mutable reference to the Notcurses context of this NcPlane.

pub fn notcurses_const<'a>(&mut self) -> &'a Notcurses

Gets an immutable reference to the Notcurses context of this NcPlane.