Type Definition libnotcurses_sys::NcScale

type NcScale = ncscale_e;

Indicates how to scale an NcVisual during rendering.

NCSCALE_NONE will apply no scaling.

NCSCALE_SCALE scales a visual to the plane's size, maintaining aspect ratio.

NCSCALE_STRETCH stretches and scales the image in an attempt to fill the entirety of the plane.

NCSCALE_NONE_HIRES like NCSCALE_NONE admitting high-res blitters.

NCSCALE_SCALE_HIRES like NCSCALE_SCALE admitting high-res blitters.

The NCSCALE_* preferences are applied only for the context of NcVisual.render. You can think of it as a pipeline:

NcVisual::fromfile() → frame → NcVisual.render() → scaling → output frame → blit

where you still have the original frame. Whereas NcVisual.resize and NcVisual.resize_noninterpolative are changing that original frame.