Type Definition libnotcurses_sys::NcVisualOptions

type NcVisualOptions = ncvisual_options;

Options struct for NcVisual

If a plane is not provided, one will be created, having the exact size necessary to display the visual (this might be smaller or larger than the rendering area). if NCVISUAL_OPTION_CHILDPLANE is provided, this will be interpreted as the parent.

A subregion of the visual can be rendered using begx, begy, lenx, and leny.

Implementations

impl NcVisualOptions

NcVisualOptions Constructors

pub fn with_plane(
plane: &mut NcPlane,
scale: NcScale,
y: NcDim,
x: NcDim,
begy: NcDim,
begx: NcDim,
leny: NcDim,
lenx: NcDim,
blitter: NcBlitter,
flags: u32,
transcolor: NcRgba
) -> Self

Specify an existing plane.

If NCVISUAL_OPTION_CHILDPLANE is used in flags then the plane is interpreted as the parent NcPlane of the new plane created for this NcVisual.

pub fn without_plane(
y: NcDim,
x: NcDim,
begy: NcDim,
begx: NcDim,
leny: NcDim,
lenx: NcDim,
blitter: NcBlitter,
flags: u32,
transcolor: u32
) -> Self

pub fn fullsize_pixel_without_plane(
y: NcDim,
x: NcDim,
leny: NcDim,
lenx: NcDim
) -> Self