Type Definition libnotcurses_sys::NotcursesOptions

type NotcursesOptions = notcurses_options;

Options struct for Notcurses

Implementations

impl NotcursesOptions

NotcursesOptions Constructors

pub const fn new() -> Self

New NotcursesOptions.

pub const fn with_margins(
top: NcDimension,
right: NcDimension,
bottom: NcDimension,
left: NcDimension
) -> Self

New NotcursesOptions, with margins.

pub const fn with_flags(flags: u64) -> Self

New NotcursesOptions, with flags.

pub const fn with_all_options(
loglevel: NcLogLevel,
margin_t: NcDimension,
margin_r: NcDimension,
margin_b: NcDimension,
margin_l: NcDimension,
flags: u64
) -> Self

New NotcursesOptions, with all the options.

Arguments

loglevel

Progressively higher log levels result in more logging to stderr. By default, nothing is printed to stderr once fullscreen service begins.

margin_t, margin_r, margin_b, margin_l

Desirable margins (top, right, bottom, left).

If all are 0 (default), we will render to the entirety of the screen. If the screen is too small, we do what we can. Absolute coordinates are relative to the rendering area ((0, 0) is always the origin of the rendering area).

flags

General flags; This is expressed as a bitfield so that future options can be added without reshaping the struct. Undefined bits must be set to 0.

NCOPTION_INHIBIT_SETLOCALE

NCOPTION_NO_ALTERNATE_SCREEN

NCOPTION_NO_FONT_CHANGES

NCOPTION_NO_QUIT_SIGHANDLERS

NCOPTION_NO_WINCH_SIGHANDLER

NCOPTION_SUPPRESS_BANNERS