Module libnotcurses_sys::widgets

All the notcurses widgets.

Constants

NCMENU_OPTION_BOTTOM
NcMenuOptions flag: Bottom row (as opposed to top row).

NCMENU_OPTION_HIDING
NcMenuOptions flag: Hide the menu when not unrolled

NCPLOT_OPTION_DETECTMAXONLY
Use domain detection only for max

NCPLOT_OPTION_EXPONENTIALD
Exponential dependent axis

NCPLOT_OPTION_LABELTICKSD
Show labels for dependent axis

NCPLOT_OPTION_NODEGRADE
Use domain detection only for max

NCPLOT_OPTION_VERTICALI
Independent axis is vertical

NCPROGBAR_OPTION_RETROGRADE
proceed left/down

NCREADER_OPTION_CURSOR
Make the terminal cursor visible across the lifetime of the ncreader, and have the ncreader manage the cursor’s placement.

NCREADER_OPTION_HORSCROLL
Enable horizontal scrolling. Virtual lines can then grow arbitrarily long.

NCREADER_OPTION_NOCMDKEYS
Disable all editing shortcuts. By default, emacs-style keys are available.

NCREADER_OPTION_VERSCROLL
Enable vertical scrolling. You can then use arbitrarily many virtual lines.

NCREEL_OPTION_CIRCULAR
is navigation circular (does moving down from the last tablet move to the first, and vice versa)? only meaningful when infinitescroll is true. if infinitescroll is false, this must be false.

NCREEL_OPTION_INFINITESCROLL
is scrolling infinite (can one move down or up forever, or is an end reached?). if true, ‘circular’ specifies how to handle the special case of an incompletely-filled reel.

Type Definitions

NcMenu
menus on the top or bottom rows

NcMenuItem
Item for NcMenu.

NcMenuOptions
Options struct for NcMenu.

NcMenuSection
Section for NcMenu.

NcMultiSelector
high-level widget for selecting items from a set

NcMultiSelectorItem
an item for NcMultiSelector

NcMultiSelectorOptions
Options structure for NcMultiSelector

NcPlotF64
A histogram, bound to an NcPlane (uses non-negative f64s)

NcPlotOptions
Options struct for NcPlotF64 or NcPlotU64

NcPlotU64
A histogram, bound to an NcPlane (uses u64s)

NcProgBar
Progress bars. They proceed linearly in any of four directions.

NcProgBarOptions
Options struct for NcProgBar

NcReader
Provides a freeform input in a (possibly multiline) region

NcReaderOptions
Options struct for NcReader

NcReel
A wheel with NcTablets on the outside.

NcReelOptions
Options struct for NcReel

NcSelector
high-level widget for selecting one item from a set

NcSelectorItem
an item for NcSelector

NcSelectorOptions
Options structur for NcSelector

NcTablet
Visual tablet for NcReel

NcTree
High-level hierarchical line-based data.

NcTreeItem
Item for NcTree.

NcTreeItemCb
An NcTreeItem callback function.

NcTreeItemCbUnsafe
An NcTreeItem callback function (unsafe).

NcTreeOptions
Options struct for NcTree.