Type Definition libnotcurses_sys::widgets::NcMenuOptions

type NcMenuOptions = ncmenu_options;

Options struct for NcMenu.

Implementations

impl NcMenuOptions

NcMenuOptions constructors

pub fn new(sections: &mut [NcMenuSection]) -> Self

New NcMenuOptions for NcMenu.

sections must contain at least 1 NcMenuSection.

pub fn with_all_args(
sections: &mut [NcMenuSection],
style_header: NcChannels,
style_sections: NcChannels,
flags: u64
) -> Self

New NcMenuOptions for NcMenu, with all args.

sections must contain at least 1 NcMenuSection.

impl NcMenuOptions

NcMenuOptions methods

pub const fn header_channels(&self) -> NcChannels

Returns the styling for the header.

(No equivalent C style function)

pub fn header_channels_mut(&mut self) -> &mut NcChannels

Returns a mutable reference of the styling for the sections.

(No equivalent C style function)

pub const fn section_channels(&self) -> NcChannels

Returns the styling for the sections.

(No equivalent C style function)

pub fn section_channels_mut(&mut self) -> &mut NcChannels

Returns a mutable reference of the styling for the sections.

(No equivalent C style function)