Type Definition libnotcurses_sys::NcBoxMask

type NcBoxMask = u32;

Controls the drawing of borders, gradients and corners.

NcBoxMax is defined in the least significant byte, where bits [3, 0] are are a border mask, and bits [7, 4] are a gradient mask.

The drawing of the corners is defined in the second byte, see NCBOXCORNER_MASK.

Diagram

NCBOXMASK_TOP    0x0001  0b00000001
NCBOXMASK_RIGHT  0x0002  0b00000010
NCBOXMASK_BOTTOM 0x0004  0b00000100
NCBOXMASK_LEFT   0x0008  0b00001000

NCBOXGRAD_TOP    0x0010  0b00010000
NCBOXGRAD_RIGHT  0x0020  0b00100000
NCBOXGRAD_BOTTOM 0x0040  0b01000000
NCBOXGRAD_LEFT   0x0080  0b10000000

NCBOXCORNER_MASK  0x0300  0b00000111_00000000

NCBOXCORNER_SHIFT 8

Bit masks

NCBOXMASK_TOP

NCBOXMASK_RIGHT

NCBOXMASK_BOTTOM

NCBOXMASK_LEFT

NCBOXGRAD_TOP

NCBOXGRAD_RIGHT

NCBOXGRAD_BOTTOM

NCBOXGRAD_LEFT

NCBOXCORNER_MASK

NCBOXCORNER_SHIFT