Type Definition libnotcurses_sys::NcDirectF

type NcDirectF = NcVisual;

A type alias of NcVisual (NcDirect Frame) intended to be used with its ncdirectf_* methods, in NcDirect mode.

Implementations

impl NcDirectF

NcDirectF Constructors & destructors

pub fn ncdirectf_from_file<'a>(
ncd: &mut NcDirect,
file: &str
) -> NcResult<&'a mut NcDirectF>

Loads media from disk, but do not yet renders it (presumably because you want to get its geometry via ncdirectf_geom(), or to use the same file with ncdirectf_render() multiple times).

You must destroy the result with ncdirectf_free();

C style function: ncdirectf_from_file().

pub fn ncdirectf_free(&mut self)

Frees a NcDirectF returned from ncdirectf_from_file().

C style function: ncdirectf_free().

impl NcDirectF

NcDirectF Methods

pub fn ncdirectf_render(
&mut self,
ncd: &mut NcDirect,
options: &NcVisualOptions
) -> NcResult<&mut NcPlane>

Same as NcDirect.render_frame(), except frame must already have been loaded.

A loaded frame may be rendered in different ways before it is destroyed.

C style function: ncvisual_render().

pub fn ncdirectf_geom(
&mut self,
ncd: &mut NcDirect,
options: &NcVisualOptions
) -> NcResult<NcVGeom>

Having loaded the frame, get the geometry of a potential render.

C style function: ncdirectf_geom().