Type Definition libnotcurses_sys::NcPalette

type NcPalette = ncpalette;

NcPalette structure consisting of an array of 256 NcChannels.

See also NcPaletteIndex.

Some terminals only support 256 colors, but allow the full palette to be specified with arbitrary RGB colors. In all cases, it's more performant to use indexed colors, since it's much less data to write to the terminal. If you can limit yourself to 256 colors, that's probably best.

type in C: ncncpalette (struct)

Implementations

impl NcPalette

pub fn new<'a>(nc: &mut Notcurses) -> &'a mut Self

New NcPalette.

C style function: ncpalette_new().

pub fn free(&mut self)

Frees this NcPalette.

C style function: ncpalette_free().

pub fn use(&self, nc: &mut Notcurses) -> NcResult<()>

Attempts to configure the terminal with this NcPalette.

C style function: ncpalette_use().

pub fn get_rgb8(&self, index: NcPaletteIndex) -> (NcColor, NcColor, NcColor)

Returns the NcColor RGB components from the NcChannel in this NcPalette.

C style function: ncpalette_get_rgb().

pub fn get_rgb(&self, index: NcPaletteIndex) -> NcChannel

Extracts the NcColor RGB components from an NcChannel entry inside this NcPalette, and returns the NcChannel.

C style function: ncpalette_get_rgb().

pub fn set(&mut self, index: NcPaletteIndex, rgb: NcRgb)

Sets the NcRgb value of the NcChannel entry inside this NcPalette.

C style function: ncpalette_set().