Type Definition libnotcurses_sys::NcVisualOptions

type NcVisualOptions = ncvisual_options;

Options struct for NcVisual

If a plane is not provided, one will be created, having the exact size necessary to display the visual.

A subregion of the visual can be rendered using begx, begy, lenx, and leny.

Implementations

impl NcVisualOptions

NcVisualOptions Constructors

pub fn with_plane(
plane: &mut NcPlane,
scale: NcScale,
y: NcDim,
x: NcDim,
begy: NcDim,
begx: NcDim,
leny: NcDim,
lenx: NcDim,
blitter: NcBlitter,
flags: u64,
transcolor: NcRgba
) -> Self

Specify an existing plane

pub fn without_plane(
y: NcDim,
x: NcDim,
begy: NcDim,
begx: NcDim,
leny: NcDim,
lenx: NcDim,
blitter: NcBlitter,
flags: u64,
transcolor: u32
) -> Self

pub fn fullsize_pixel_without_plane(
y: NcDim,
x: NcDim,
leny: NcDim,
lenx: NcDim
) -> Self