Type Definition libnotcurses_sys::Notcurses

type Notcurses = notcurses;

The main struct of the (full mode) TUI library

Notcurses builds atop the terminfo abstraction layer to provide reasonably portable vivid character displays.

Implementations

impl Notcurses

Notcurses Constructors

pub fn new<'a>() -> &'a mut Notcurses

Returns a Notcurses context (without banners).

pub fn with_banners<'a>() -> &'a mut Notcurses

Returns a Notcurses context, with banners. The default in the C library.

pub fn without_altscreen<'a>() -> &'a mut Notcurses

Returns a Notcurses context, without an alternate screen (nor banners).

pub fn without_altscreen_nor_banners<'a>() -> &'a mut Notcurses

Returns a Notcurses context, without an alternate screen, with banners.

pub fn with_options<'a>(options: &NotcursesOptions) -> &'a mut Notcurses

Returns a Notcurses context, expects NotcursesOptions.

impl Notcurses

Notcurses methods

pub fn stdplane<'a>(&mut self) -> &'a mut NcPlane

Returns a mutable reference to the standard NcPlane for this terminal.

The standard plane always exists, and its origin is always at the uppermost, leftmost cell.

pub fn stdplane_const<'a>(&self) -> &'a NcPlane

Returns a reference to the standard NcPlane for this terminal.

The standard plane always exists, and its origin is always at the uppermost, leftmost cell.

pub fn stop(&mut self) -> NcResult

Destroy the Notcurses context.

pub fn render(&mut self) -> NcResult